Edgewater correctly clips so the H2 can be viewed.   This is taken from ButterflyEngMerge but simplified

+++ System/attachCompInc/LayerComp.h
@@ -46,7 +46,7 @@ class LayerComp
   LayerComp& operator=(const LayerComp&);
   virtual ~LayerComp() { }  ///< Simple destructor
 
-  virtual double intersectDist(const size_t,const size_t,
+  virtual double intersectDist(const size_t,const long int,
 			       const Geometry::Vec3D&,
 			       const Geometry::Vec3D&) const;
